+++ src/modules/leibnitz-codegen.ts
@@ -1,3 +1,27 @@
+/*
+MIT License
+
+Copyright (c) 2018 Marco Marini
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+*/
+
 import { Quaternion } from '@babylonjs/core';
 import _ from 'lodash';
 import {
@@ -28,7 +52,8 @@ import {
     createScalarMulCode,
     createQuaternionRefCode,
     createMatrixRefCode,
-    createModScalarCode
+    createModScalarCode,
+    createMulQMCode
 } from "./leibniz-code";
 import { ValueFunction } from "./leibniz-defs";
 import { ASTNode, ASTNodeType, baseVectorId, kroneckerId, TreeNode } from "./leibniz-parser";
@@ -711,6 +736,11 @@ function mulCodeGen(ctx: CodeGenContext, a: ValueCode, b: ValueCode): CodeGenCon
             case ValueTypeCode.Quaternion:
                 return ctx.withTypeCode(newQuaternionCode(createMulQQCode(a.code, b.code)));
             case ValueTypeCode.Vector:
+                if (b.rows === 3) {
+                    return ctx.withTypeCode(newVectorCode(createMulQMCode(a.code, b.code), 3));
+                } else {
+                    return ctx.addErrors(errorByCode('vector3 expected', b));
+                }
             case ValueTypeCode.Matrix:
                 return ctx.addErrors(errorByCode('scalar or quaternion expected', b));
         }