+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Issue #65: Add product quaternion by vector3
+
 ## [2.0.1] - 2023-09-07
 
 ### Added